+++ packages/amplify-e2e-tests/src/__tests__/cdk/base-cdk.test.ts
@@ -1,6 +1,6 @@
 import * as path from 'path';
 import { createNewProjectDir, deleteProjectDir, initCDKProject, cdkDeploy, cdkDestroy } from 'amplify-category-api-e2e-core';
-import { default as fetch, Request } from 'node-fetch';
+import { graphql } from './graphql-utils';
 
 describe('CDK GraphQL Transformer', () => {
   let projRoot: string;
@@ -15,7 +15,7 @@ describe('CDK GraphQL Transformer', () => {
     try {
       await cdkDestroy(projRoot, '--all');
     } catch (_) {
-      // No-op.
+      /* No-op */
     }
 
     deleteProjectDir(projRoot);
@@ -26,20 +26,6 @@ describe('CDK GraphQL Transformer', () => {
     const name = await initCDKProject(projRoot, templatePath);
     const outputs = await cdkDeploy(projRoot, '--all');
     const { GraphQLAPIEndpointOutput: apiEndpoint, GraphQLAPIKeyOutput: apiKey } = outputs[name];
-
-    const assertGraphQLQuerySnapshot = async (query: string, queryName: string) => {
-      const result = await graphql(apiEndpoint, apiKey, query);
-      expect(result).toMatchSnapshot({
-        body: {
-          data: {
-            [queryName]: {
-              id: expect.any(String),
-            },
-          },
-        },
-      });
-      return result;
-    };
 
     const result = await graphql(
       apiEndpoint,
@@ -53,6 +39,7 @@ describe('CDK GraphQL Transformer', () => {
         }
       `,
     );
+
     expect(result).toMatchSnapshot({
       body: {
         data: {
@@ -96,42 +83,3 @@ describe('CDK GraphQL Transformer', () => {
     expect(todo.id).toEqual(listResult.body.data.listTodos.items[0].id);
   });
 });
-
-async function graphql(apiEndpoint: string, apiKey: string, query: string) {
-  const options = {
-    method: 'POST',
-    headers: {
-      'x-api-key': apiKey,
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify({ query }),
-  };
-
-  const request = new Request(apiEndpoint, options);
-
-  let statusCode = 200;
-  let body;
-  let response;
-
-  try {
-    response = await fetch(request);
-    body = await response.json();
-    if (body.errors) statusCode = 400;
-  } catch (error) {
-    statusCode = 400;
-    body = {
-      errors: [
-        {
-          status: response?.status,
-          message: error.message,
-          stack: error.stack,
-        },
-      ],
-    };
-  }
-
-  return {
-    statusCode,
-    body: body,
-  };
-}