+++ packages/amplify-e2e-tests/src/init-special-cases/index.ts
@@ -30,8 +30,6 @@ export async function initWithoutCredentialFileAndNoNewUserSetup(projRoot) {
       fs.renameSync(credentialsFilePath, credentialsFilePathHide);
     }
     await initWorkflow(projRoot, settings);
-  } catch (e) {
-    throw e;
   } finally {
     if (fs.existsSync(configFilePathHide)) {
       fs.renameSync(configFilePathHide, configFilePath);