+++ packages/amplify-graphql-model-transformer/src/graphql-model-transformer.ts
@@ -698,7 +698,7 @@ export class ModelTransformer extends TransformerModelBase implements Transforme
       }
       case SubscriptionFieldType.ON_CREATE:
       case SubscriptionFieldType.ON_DELETE:
-      case SubscriptionFieldType.ON_UPDATE:
+      case SubscriptionFieldType.ON_UPDATE: {
         const filterInputName = toPascalCase(['ModelSubscription', type.name.value, 'FilterInput']);
         const filterInputs = createEnumModelFilters(ctx, type);
         filterInputs.push(makeSubscriptionQueryFilterInput(ctx, filterInputName, type));
@@ -710,6 +710,7 @@ export class ModelTransformer extends TransformerModelBase implements Transforme
         });
 
         return [makeInputValueDefinition('filter', makeNamedType(filterInputName))];
+      }
 
       default:
         throw new Error('Unknown operation type');