+++ packages/amplify-graphql-transformer-core/src/graphql-api.ts
@@ -9,8 +9,10 @@ import {
   OpenIdConnectConfig,
   UserPoolConfig,
   UserPoolDefaultAction,
+  CfnApiKey,
+  CfnGraphQLApi,
+  CfnGraphQLSchema,
 } from 'aws-cdk-lib/aws-appsync';
-import { CfnApiKey, CfnGraphQLApi, CfnGraphQLSchema } from 'aws-cdk-lib/aws-appsync';
 import { Grant, IGrantable, ManagedPolicy, Role, ServicePrincipal } from 'aws-cdk-lib/aws-iam';
 import * as cdk from 'aws-cdk-lib';
 import { ArnFormat, CfnResource, Duration, Stack } from 'aws-cdk-lib';