+++ packages/graphql-transformers-e2e-tests/src/__tests__/ConnectionsWithAuthTests.e2e.test.ts
@@ -5,8 +5,7 @@ import { DynamoDBModelTransformer } from 'graphql-dynamodb-transformer';
 import { ModelAuthTransformer } from 'graphql-auth-transformer';
 import { ModelConnectionTransformer } from 'graphql-connection-transformer';
 import { Output } from 'aws-sdk/clients/cloudformation';
-import { default as S3 } from 'aws-sdk/clients/s3';
-import { CreateBucketRequest } from 'aws-sdk/clients/s3';
+import { default as S3, CreateBucketRequest } from 'aws-sdk/clients/s3';
 import { default as CognitoClient } from 'aws-sdk/clients/cognitoidentityserviceprovider';
 import { default as moment } from 'moment';
 import { GraphQLClient } from '../GraphQLClient';
@@ -248,7 +247,7 @@ afterAll(async () => {
 /**
  * Tests
  */
-test('Test creating a post and immediately view it via the User.posts connection.', async () => {
+test('creating a post and immediately view it via the User.posts connection.', async () => {
   const createUser1 = await GRAPHQL_CLIENT_1.query(
     `mutation {
         createUser(input: { id: "user1@test.com" }) {
@@ -339,7 +338,7 @@ test('Testing reading an owner protected field as a non owner', async () => {
   expect(response3.data.getFieldProtected.ownerOnly).toEqual('owner-protected');
 });
 
-test('Test that @connection resolvers respect @model read operations.', async () => {
+test('that @connection resolvers respect @model read operations.', async () => {
   const response1 = await GRAPHQL_CLIENT_1.query(
     `mutation {
         createOpenTopLevel(input: { id: "1", owner: "${USERNAME1}", name: "open" }) {
@@ -406,7 +405,7 @@ test('Test that @connection resolvers respect @model read operations.', async ()
 });
 
 // Per field auth in mutations
-test('Test that owners cannot set the field of a FieldProtected object unless authorized.', async () => {
+test('that owners cannot set the field of a FieldProtected object unless authorized.', async () => {
   const response1 = await GRAPHQL_CLIENT_1.query(
     `mutation {
         createFieldProtected(input: { id: "2", owner: "${USERNAME1}", ownerOnly: "owner-protected" }) {
@@ -454,7 +453,7 @@ test('Test that owners cannot set the field of a FieldProtected object unless au
   expect(response3.errors).toHaveLength(1);
 });
 
-test('Test that owners cannot update the field of a FieldProtected object unless authorized.', async () => {
+test('that owners cannot update the field of a FieldProtected object unless authorized.', async () => {
   const response1 = await GRAPHQL_CLIENT_1.query(
     `mutation {
         createFieldProtected(input: { owner: "${USERNAME1}", ownerOnly: "owner-protected" }) {