+++ packages/amplify-migration-tests/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.35](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-migration-tests@5.0.34...amplify-category-api-migration-tests@5.0.35) (2023-08-30)
+
+**Note:** Version bump only for package amplify-category-api-migration-tests
+
+## [5.0.34](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-migration-tests@5.0.33...amplify-category-api-migration-tests@5.0.34) (2023-08-28)
+
+**Note:** Version bump only for package amplify-category-api-migration-tests
+
 ## [5.0.33](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-migration-tests@5.0.32...amplify-category-api-migration-tests@5.0.33) (2023-08-09)
 
 **Note:** Version bump only for package amplify-category-api-migration-tests