+++ packages/graphql-function-transformer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.56](https://github.com/aws-amplify/amplify-category-api/compare/graphql-function-transformer@3.3.55...graphql-function-transformer@3.3.56) (2023-08-30)
+
+**Note:** Version bump only for package graphql-function-transformer
+
+## [3.3.55](https://github.com/aws-amplify/amplify-category-api/compare/graphql-function-transformer@3.3.54...graphql-function-transformer@3.3.55) (2023-08-28)
+
+**Note:** Version bump only for package graphql-function-transformer
+
 ## [3.3.54](https://github.com/aws-amplify/amplify-category-api/compare/graphql-function-transformer@3.3.53...graphql-function-transformer@3.3.54) (2023-08-09)
 
 **Note:** Version bump only for package graphql-function-transformer